Move Model related helper fucntions to models_utils.py
