+++ forge/test/models/models_utils.py
@@ -6,6 +6,9 @@ import requests
 from PIL import Image
 from torchvision import transforms
 from transformers import AutoImageProcessor
+import torch
+from tabulate import tabulate
+import json
 
 # Mean Pooling - Take attention mask into account for correct averaging
 def mean_pooling(model_output, attention_mask):
@@ -61,3 +64,40 @@ def get_sample_data(model_name):
     image_processor = AutoImageProcessor.from_pretrained(model_name)
     pixel_values = image_processor(images=image, return_tensors="pt").pixel_values
     return [pixel_values]
+
+
+imagenet_class_index_path = "forge/test/models/files/labels/imagenet_class_index.json"
+
+
+def load_class_labels(file_path):
+    """Load class labels from the local JSON file."""
+    with open(file_path, "r") as f:
+        class_idx = json.load(f)
+    return [class_idx[str(i)][1] for i in range(len(class_idx))]
+
+
+def print_cls_results(fw_out, compiled_model_out):
+
+    class_labels = load_class_labels(imagenet_class_index_path)
+    fw_top1_probabilities, fw_top1_class_indices = torch.topk(fw_out.softmax(dim=1) * 100, k=1)
+    compiled_model_top1_probabilities, compiled_model_top1_class_indices = torch.topk(
+        compiled_model_out.softmax(dim=1) * 100, k=1
+    )
+
+    # Directly get the top 1 predicted class and its probability for both models
+    fw_top1_class_idx = fw_top1_class_indices[0, 0].item()
+    compiled_model_top1_class_idx = compiled_model_top1_class_indices[0, 0].item()
+    fw_top1_class_prob = fw_top1_probabilities[0, 0].item()
+    compiled_model_top1_class_prob = compiled_model_top1_probabilities[0, 0].item()
+
+    # Get the class labels for top 1 class
+    fw_top1_class_label = class_labels[fw_top1_class_idx]
+    compiled_model_top1_class_label = class_labels[compiled_model_top1_class_idx]
+
+    # Prepare the results for displaying
+    table = [
+        ["Metric", "Framework Model", "Compiled Model"],
+        ["Top 1 Predicted Class Label", fw_top1_class_label, compiled_model_top1_class_label],
+        ["Top 1 Predicted Class Probability", fw_top1_class_prob, compiled_model_top1_class_prob],
+    ]
+    print(tabulate(table, headers="firstrow", tablefmt="grid"))
